+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [1.5.16] - 2025-12-18
 
 ### Added
-- Added timestamp toggle in sidebar to show/hide static timestamps alongside GitHub's relative timestamps
+- Added timestamp toggle in sidebar to show/hide absolute timestamps alongside GitHub's relative timestamps
 
 Check out the K2 repo on GitHub to see what's new!